use double instead of float
git-svn-id: file:///srv/svnroot/semilib/trunk@1713 a19b6091-82f5-0310-b5c5-f4b00232789b

+++ fuzzy/SoundexFunctor.cpp
@@ -40,7 +40,7 @@ Ranking Soundex::operator () ( const string & element )
 	string soundex = PrepareTheString ( element );
 
 	// return percentage match and original string
-	float percent = (( soundex == _toMatch ) ? 1.0 : 0.0 );
+	double percent = (( soundex == _toMatch ) ? 1.0 : 0.0 );
 	return Ranking ( percent, element );
 }
